+++ migrations/2021-11-22-030926_create_visited_cities/down.sql
@@ -1,0 +1,1 @@
+drop table visited_cities;