+++ migrations/2021-11-22-030926_create_visited_cities/up.sql
@@ -1,0 +1,6 @@
+CREATE TABLE visited_cities
+(
+    user_id int not null,
+    city_id int not null,
+    primary key (user_id, city_id)
+);